(PC-16477) refactor(profile): add new domains_credit fixture
