+++ src/features/cheatcodes/pages/NavigationProfile/NavigationProfile.tsx
@@ -3,11 +3,11 @@ import { ScrollView } from 'react-native'
 import styled from 'styled-components/native'
 
 import { LinkToComponent } from 'features/cheatcodes/components/LinkToComponent'
+import { useGoBack } from 'features/navigation/useGoBack'
 import {
   domains_credit_v1,
   domains_credit_v2,
-} from 'features/cheatcodes/pages/AppComponents/AppComponents'
-import { useGoBack } from 'features/navigation/useGoBack'
+} from 'features/profile/components/headers/fixtures/domainsCredit'
 import { CreditCeilingsModal } from 'features/profile/components/headers/Modals/CreditCeilingsModal'
 import { ExpiredCreditModal } from 'features/profile/components/headers/Modals/ExpiredCreditModal'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
