+++ src/features/profile/components/BeneficiaryCeilings.stories.tsx
@@ -1,6 +1,14 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 import React from 'react'
 
+import {
+  domains_credit_underage,
+  domains_credit_v1,
+  domains_credit_v2,
+  domains_exhausted_credit_underage,
+  domains_exhausted_credit_v1,
+} from 'features/profile/components/headers/fixtures/domainsCredit'
+
 import { BeneficiaryCeilings } from './BeneficiaryCeilings'
 
 export default {
@@ -11,31 +19,6 @@ export default {
 const Template: ComponentStory<typeof BeneficiaryCeilings> = (args) => (
   <BeneficiaryCeilings {...args} />
 )
-
-const domains_credit_v1 = {
-  all: { initial: 50000, remaining: 40000 },
-  physical: { initial: 30000, remaining: 10000 },
-  digital: { initial: 30000, remaining: 20000 },
-}
-
-const domains_exhausted_credit_v1 = {
-  all: { initial: 50000, remaining: 0 },
-  physical: { initial: 30000, remaining: 0 },
-  digital: { initial: 30000, remaining: 0 },
-}
-
-const domains_credit_v2 = {
-  all: { initial: 30000, remaining: 10000 },
-  digital: { initial: 20000, remaining: 5000 },
-}
-
-const domains_credit_underage = {
-  all: { initial: 3000, remaining: 1000 },
-}
-
-const domains_exhausted_credit_underage = {
-  all: { initial: 3000, remaining: 0 },
-}
 
 export const Default = Template.bind({})
 Default.args = {
