+++ src/locales/fr/messages.po
@@ -1770,6 +1770,7 @@ msgstr "Mois suivant"
 
 #: src/features/profile/components/ExBeneficiaryHeader.tsx
 #: src/features/profile/components/ExBeneficiaryHeader.tsx
+#: src/features/profile/components/ExhaustedCreditModal.tsx
 #: src/features/profile/components/headers/Modals/ExpiredCreditModal.tsx
 msgid "Mon crédit est expiré, que faire ?"
 msgstr "Mon crédit est expiré, que faire ?"
@@ -3199,6 +3200,10 @@ msgstr "Tu peux activer ou désactiver cette fonctionnalité dans les paramètre
 msgid "Tu peux aussi compléter ton dossier sur Démarches simplifiées. Attention le traitement sera plus long !"
 msgstr "Tu peux aussi compléter ton dossier sur Démarches simplifiées. Attention le traitement sera plus long !"
 
+#: src/features/profile/components/ExhaustedCreditModal.tsx
+msgid "Tu peux aussi découvrir les activités culturelles de nos partenaires sur l'application et les réserver directement sur leur site."
+msgstr "Tu peux aussi découvrir les activités culturelles de nos partenaires sur l'application et les réserver directement sur leur site."
+
 #: src/features/profile/components/headers/Modals/ExpiredCreditModal.tsx
 msgid "Tu peux aussi découvrir les activités culturelles de nos partenaires sur l'application et réserver directement sur leur site."
 msgstr "Tu peux aussi découvrir les activités culturelles de nos partenaires sur l'application et réserver directement sur leur site."
@@ -3243,6 +3248,7 @@ msgstr "Tu peux sélectionner une ou plusieurs réponses."
 msgid "Tu peux toujours réserver les offres gratuites exclusives au pass Culture."
 msgstr "Tu peux toujours réserver les offres gratuites exclusives au pass Culture."
 
+#: src/features/profile/components/ExhaustedCreditModal.tsx
 #: src/features/profile/components/headers/Modals/ExpiredCreditModal.tsx
 msgid "Tu peux toujours réserver les offres gratuites ou exclusives sur le pass Culture."
 msgstr "Tu peux toujours réserver les offres gratuites ou exclusives sur le pass Culture."
